feat: Add ComponentCodeValidator class
This commit adds the `ComponentCodeValidator` class to the `helpers` module in the `langflow/components` package. The `ComponentCodeValidator` class is responsible for validating the code of a component. It takes a `component_code` input and uses it to build a custom component template. If there are any errors during the validation process, the error message and traceback are returned. This new class enhances the functionality of the application by providing code validation capabilities.